Added CONTRIBUTING.md and modified README.md (#17)
I used the [PyBaMM's
CONTRIBUTING.md](https://github.com/pybamm-team/PyBaMM/blob/develop/CONTRIBUTING.md)
file as a template for this project as well. It was modified and omitted
for it to be in accordance with this project. `README.md` file had to be
written mostly on its own.

The parameter sets have to be included
within the template in order for them to get copied to the
user-generated cookiecutter project, if you want to use the parameter
sets feature I suggest you do an editable install to the
`pybamm-cookiecutter` project under a virtual environment, the
instructions are in `README.md`.

---------

Co-authored-by: Agriya Khetarpal <74401230+agriyakhetarpal@users.noreply.github.com>
Co-authored-by: Arjun Verma <arjunverma.oc@gmail.com>